Updated star rating on each review tile to show remaining empty stars
