+++ client/src/assets/ratings-and-reviews.sass
@@ -166,6 +166,11 @@
               flex-direction: row
               justify-content: flex-start
 
+            #reviewCharacteristics
+              display: flex
+              flex-display: row
+              justify-content: space-between
+
             .reviewInput
               text-align: left
