+++ client/src/components/RatingsAndReviews/RatingBreakdown.jsx
@@ -2,10 +2,10 @@ import React, { useState, useEffect } from 'react';
 import StarRating from './StarRating.jsx';
 import { getAverageRating, calculateTotalReviews, calculateRatingsPercentage, getRecommendPercentage } from '../../lib/ratingsAndReviewsHelpers.js';
 import { getReviewsMeta } from '../../lib/requestHelpers.js';
-import { FaStar, FaStarHalfAlt, FaRegStar } from 'react-icons/fa';
 
 
-const RatingBreakdown = ({ metaData, reveiws, setReviews, sortReviews, activeStars, removeAllFilters }) => {
+
+const RatingBreakdown = ({ metaData, reveiws, setReviews, sortReviews, activeStars, removeAllFilters, convertRatingToStars }) => {
 
   const [filterActive, setFilterActive] = useState(false);
 
@@ -14,36 +14,6 @@ const RatingBreakdown = ({ metaData, reveiws, setReviews, sortReviews, activeSta
   let averageRating = getAverageRating(ratings, 1);
   let percentage = calculateRatingsPercentage(ratings, totalReviews);
   let recommendPercentage = getRecommendPercentage(metaData.recommended);
-
-  const convertRatingToStars = (rating) => {
-    let fullStars = Math.floor(rating);
-    let remainder = 5 - rating;
-    let emptyStars = Math.floor(remainder);
-    let partial = remainder - emptyStars;
-
-    const renderStar = (type) =>{
-      if (type === 'full') {
-        return <FaStar />;
-      } else if (type === 'half') {
-        return <FaStarHalfAlt />;
-      } else if (type === 'empty') {
-        return <FaRegStar />;
-      }
-    };
-
-    let stars = [];
-
-    for (let i = 0; i < fullStars; i ++) {
-      stars.push(renderStar('full'));
-    }
-    if (partial) {
-      stars.push(renderStar('half'));
-    }
-    for (let j = 0; j < emptyStars; j ++) {
-      stars.push(renderStar('empty'));
-    }
-    return stars;
-  };
 
   return (
     <div id="breakdown">
