+++ client/src/components/RatingsAndReviews/RatingsAndReviews.jsx
@@ -6,6 +6,7 @@ import AddReview from './AddReview.jsx';
 import Characteristics from './Characteristics.jsx';
 import AddReviewModal from './AddReviewModal.jsx';
 import { getReviews, getReviewsMeta } from '../../lib/requestHelpers.js';
+import { FaStar, FaStarHalfAlt, FaRegStar } from 'react-icons/fa';
 
 
 const RatingsAndReviews = ({ product }) => {
@@ -15,6 +16,11 @@ const RatingsAndReviews = ({ product }) => {
   const [visibleReviews, setVisibleReviews] = useState([]);
   const [isOpen, setIsOpen] = useState(false);
   const [activeStars, setActiveStars] = useState([]);
+
+  const returnReviewsMeta = () => {
+    return reviewsMeta;
+  };
+
 
   useEffect(() => {
     getReviews(product)
@@ -66,6 +72,36 @@ const RatingsAndReviews = ({ product }) => {
     setVisibleReviews(reviews.slice(0, 2));
   };
 
+  const convertRatingToStars = (rating) => {
+    let fullStars = Math.floor(rating);
+    let remainder = 5 - rating;
+    let emptyStars = Math.floor(remainder);
+    let partial = remainder - emptyStars;
+
+    const renderStar = (type) =>{
+      if (type === 'full') {
+        return <FaStar />;
+      } else if (type === 'half') {
+        return <FaStarHalfAlt />;
+      } else if (type === 'empty') {
+        return <FaRegStar />;
+      }
+    };
+
+    let stars = [];
+
+    for (let i = 0; i < fullStars; i ++) {
+      stars.push(renderStar('full'));
+    }
+    if (partial) {
+      stars.push(renderStar('half'));
+    }
+    for (let j = 0; j < emptyStars; j ++) {
+      stars.push(renderStar('empty'));
+    }
+    return stars;
+  };
+
   return (
     <>
       <div id="ratingsAndReviews">
@@ -75,20 +111,20 @@ const RatingsAndReviews = ({ product }) => {
         <section id="breakdownAndReviews">
           {reviews.length > 0 &&
             <div id="breakdownAndCharacteristics">
-              {reviewsMeta && <RatingBreakdown sortReviews={sortReviews} metaData={reviewsMeta} reviews={reviews} setReviews={setReviews} activeStars={activeStars} removeAllFilters={removeAllFilters}/>}
+              {reviewsMeta && <RatingBreakdown convertRatingToStars={convertRatingToStars} sortReviews={sortReviews} metaData={reviewsMeta} reviews={reviews} setReviews={setReviews} activeStars={activeStars} removeAllFilters={removeAllFilters}/>}
               {reviewsMeta && <Characteristics metaData={reviewsMeta} />}
             </div>}
 
           <div id="reviewsListAndButtons">
             <div id="listSortHeading">{reviews.length} reviews, sorted by relevance</div>
-            {reviews.length > 0 && <ReviewsList visibleReviews={visibleReviews} setVisibleReviews={setVisibleReviews} reviews={reviews} />}
+            {reviews.length > 0 && <ReviewsList convertRatingToStars={convertRatingToStars} visibleReviews={visibleReviews} setVisibleReviews={setVisibleReviews} reviews={reviews} />}
 
             <div id="bottomButtons">
               {reviews.length !== visibleReviews.length && <button className="reviewButton" onClick={addReviews}>MORE REVIEWS</button>}
               {reviewsMeta && !isOpen && <button className="reviewButton" onClick={() => {
                 setIsOpen(true);
               }}>ADD REVIEW +</button>}
-              <AddReviewModal product={product} metaData={reviewsMeta} open={isOpen} onClose={() => {
+              <AddReviewModal returnReviewsMeta={returnReviewsMeta} product={product} metaData={reviewsMeta} open={isOpen} onClose={() => {
                 setIsOpen(false);
               }} />
             </div>
