+++ client/src/components/RatingsAndReviews/ReviewsList.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import AddReview from './AddReview.jsx';
 import ReviewTile from './ReviewTile.jsx';
 
-const ReviewsList = ({reviews, visibleReviews, setVisibleReviews}) => {
+const ReviewsList = ({reviews, visibleReviews, setVisibleReviews, convertRatingToStars }) => {
 
   useEffect(() => {
     setVisibleReviews(reviews.slice(0, 2));
@@ -13,7 +13,7 @@ const ReviewsList = ({reviews, visibleReviews, setVisibleReviews}) => {
   return (
     <div id="reviewsList">
       {visibleReviews.map((review) =>
-        <ReviewTile review={review} key={review.review_id} />
+        <ReviewTile review={review} key={review.review_id} convertRatingToStars={convertRatingToStars}/>
       )}
     </div>
   );
